# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Memory Atlas API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/session": {
      "post": {
        "summary": "Create Session",
        "operationId": "create_session_api_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSession"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureSession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/session/{session_id}": {
      "get": {
        "summary": "Get Session",
        "operationId": "get_session_api_session__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureSession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Session",
        "operationId": "update_session_api_session__session_id__patch",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionPreferences"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureSession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ingest": {
      "post": {
        "summary": "Ingest",
        "operationId": "ingest_api_ingest_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_api_ingest_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureSession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/source/{source_id}": {
      "get": {
        "summary": "Source",
        "operationId": "source_api_source__source_id__get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceFragment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Source",
        "operationId": "update_source_api_source__source_id__patch",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FragmentPreferences"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceFragment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Source",
        "operationId": "delete_source_api_source__source_id__delete",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureSession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/source/{source_id}/media": {
      "get": {
        "summary": "Source Media",
        "operationId": "source_media_api_source__source_id__media_get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/source/{source_id}/text": {
      "put": {
        "summary": "Edit Text",
        "operationId": "edit_text_api_source__source_id__text_put",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceFragment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/session/{session_id}/conversation": {
      "post": {
        "summary": "Add Conversation",
        "operationId": "add_conversation_api_session__session_id__conversation_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ConversationTurn"
                },
                "title": "Turns"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureSession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_ingest_api_ingest_post": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files",
            "default": []
          },
          "text_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Json"
          }
        },
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "Body_ingest_api_ingest_post"
      },
...
      "CaptureSession": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "0.1.0",
            "title": "Schema Version",
            "default": "0.1.0"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "allow_text_refinement": {
            "type": "boolean",
            "title": "Allow Text Refinement",
            "default": false
          },
          "weave_instruction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weave Instruction",
            "description": "Author's optional free-text guidance for the weaver (e.g. 'focus on Shelly')."
          },
          "woven": {
            "type": "boolean",
            "title": "Woven",
            "description": "True once a weave has finished for this session; Capture then starts a fresh notebook page.",
            "default": false
          },
          "conversation": {
            "items": {
              "$ref": "#/components/schemas/ConversationTurn"
            },
            "type": "array",
            "title": "Conversation",
            "default": []
          },
          "fragments": {
            "items": {
              "$ref": "#/components/schemas/SourceFragment"
            },
            "type": "array",
            "title": "Fragments"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "created_at",
          "fragments"
        ],
        "title": "CaptureSession"
      },
...
      "ConversationTurn": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "weaver",
              "author"
            ],
            "title": "Role"
          },
          "text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "question_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Id"
          },
          "at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "role",
          "text"
        ],
        "title": "ConversationTurn"
      },
      "CreateSession": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CreateSession"
      },
...
      "FragmentPreferences": {
        "properties": {
          "keep_original": {
            "type": "boolean",
            "title": "Keep Original"
          },
          "source_instruction": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Instruction"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "keep_original"
        ],
        "title": "FragmentPreferences"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessionPreferences": {
        "properties": {
          "allow_text_refinement": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allow Text Refinement"
          },
          "weave_instruction": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Weave Instruction"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SessionPreferences"
      },
      "SourceFragment": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "0.1.0",
            "title": "Schema Version",
            "default": "0.1.0"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "modality": {
            "type": "string",
            "enum": [
              "image",
              "text",
              "audio"
            ],
            "title": "Modality"
          },
          "original_media_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Media Ref"
          },
          "original_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Text"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "media_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Type"
          },
          "byte_size": {
            "type": "integer",
            "title": "Byte Size",
            "default": 0
          },
          "captured_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At"
          },
          "keep_original": {
            "type": "boolean",
            "title": "Keep Original",
            "default": true
          },
          "source_instruction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Instruction"
          },
          "transcript": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceTranscript"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "session_id",
          "modality"
        ],
        "title": "SourceFragment"
      },
      "SourceTranscript": {
        "properties": {
          "duration_ms": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Duration Ms"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/TranscriptSegment"
            },
            "type": "array",
            "title": "Segments"
          },
          "schema_version": {
            "type": "string",
            "const": "0.1.0",
            "title": "Schema Version",
            "default": "0.1.0"
          },
          "source_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 1,
            "minItems": 1,
            "title": "Source Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "duration_ms",
          "segments",
          "source_ids"
        ],
        "title": "SourceTranscript"
      },
...
      "TextEdit": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "TextEdit"
      },
...
      "TranscriptSegment": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "clip_start_ms": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Clip Start Ms"
          },
          "clip_end_ms": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Clip End Ms"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "clip_start_ms",
          "clip_end_ms"
        ],
        "title": "TranscriptSegment"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}